feat: Highlight checked kings with red dropshadow
This adds a red dropshadow effect to any king on the move they get checked.
Eventually, a new flag is needed to disable this feature, as well as customizing
the color. Moreover, I would like to highlight with a different dropshadow
(green maybe) the king that wins the game. SVG files are included to make this work.