feat: Add termination circles on the last frame of the game
Indicate the reason why the game ended by drawing a small circle on top of each king.
The winning king gets a crown, the loser king gets a symbol indicating the reason why
it lost (timeout, resignation, or checkmate), and in the event of a draw a 1/2 is drawn
on both kings.

This required moving all svgs under a single directory svgs/.

Finally, also updated the README with more examples.