fix: Correct flipping
When drawing a checked king piece image would appear flipped. This was
caused by not accounting for the board already being flipped when the
checked king was drawn, as this happens after the move has been
drawn. This issue would only occur when running with --flip.

A similar issue was causing termination circles to not be properly
drawn.

I've moved some responsibility to the PieceInBoard struct and
eventually I'd like to do more of that.